refactor: split subscription tests
